pyproject.toml
[tool.poetry]
name = "home_work_28"
version = "0.1.0"
description = ""
authors = ["Vasiliy <zamotaev.v-a@yandex.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
psycopg2 = "^2.9.4"
pillow = "^9.2.0"
pylint = "^2.15.5"
requests = "^2.28.1"
faker = "^15.1.1"
django-debug-toolbar = "^3.7.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-pytest = "^0.2.0"
pytest = "^7.2.0"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
